+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Improve detection for cursor agents
+- Add support for pi coding agent
+
 ## 0.1.1
 
 - Improve detection for cursor usage to avoid mistaking humans as agents